API

MixedAnova.AnovaResult

MixedAnova.FTest

MixedAnova.GoodnessOfFit

MixedAnova.LRT

MixedAnova.LikelihoodRatioTest

MixedAnova.anova

MixedAnova.anova_glm

MixedAnova.anova_lfe

MixedAnova.anova_lm

MixedAnova.anova_lme

MixedAnova.anova_test

MixedAnova.anova_type

MixedAnova.calcdof

MixedAnova.canonicalgoodnessoffit

MixedAnova.formula

MixedAnova.glme

MixedAnova.lfe

MixedAnova.lme

MixedAnova.nestedmodels

MixedAnova.pval

MixedAnova.teststat

MixedAnova.to_trm

Type

MixedAnova.AnovaResult — Type

AnovaResult{M, T, N}

Returned object of anova.

M is a subtype of Tuple if multiple models are provided; otherwise, a typeof model.

T is a subtype of GoodnessOfFit; either FTest or LRT.

N is the length of parameters.

Fields

model: full model or tuple of tested models.

type: type of anova.

dof: degree of freedoms of models or factors.

deviance: deviance(s) for calculating test statistics. See deviance for more details.

teststat: value(s) of test statiscics.

pval: p-value(s) of test statiscics.

tests: NamedTuple contained extra statistics.

MixedAnova.FTest — Type

struct FTest <: GoodnessOfFit end

ANOVA by F-test. It can be the first argument or keyword argument test.

MixedAnova.GoodnessOfFit — Type

abstract type GoodnessOfFit end

MixedAnova.LRT — Type

const LRT = LikelihoodRatioTest

See LikelihoodRatioTest(@ref).

MixedAnova.LikelihoodRatioTest — Type

struct LikelihoodRatioTest <: GoodnessOfFit end
const LRT = LikelihoodRatioTest

ANOVA by Likelihood-ratio test. It can be the first argument or keyword argument test.

ANOVA

MixedAnova.anova — Function

anova(<models>...; test::Type{<: GoodnessOfFit})

Analysis of variance.

Return AnovaResult{M, test, N}. See AnovaResult for details.

models: model objects

TableRegressionModel{<: LinearModel} fitted by GLM.lm

TableRegressionModel{<: GeneralizedLinearModel} fitted by GLM.glm

LinearMixedModel fitted by MixedAnova.lme or fit(LinearMixedModel, ...)

GeneralizedLinearMixedModel fitted by MixedAnova.glme or fit(GeneralizedLinearMixedModel, ...)

TableRegressionModel{<: FixedEffectModel} fitted by MixedAnova.lfe.

If mutiple models are provided, they should be nested and the last one is the most saturated.

test: test statistics for goodness of fit. Available tests are LikelihoodRatioTest (LRT) and FTest. The default is based on the model type.

TableRegressionModel{<: LinearModel}: FTest.

TableRegressionModel{<: GeneralizedLinearModel}: based on distribution function, see canonicalgoodnessoffit.

LinearMixedModel: FTest for one model fit; LRT for nested models.

GeneralizedLinearMixedModel: LRT for nested models.

TableRegressionModel{<: FixedEffectModel}: FTest.

When multiple models are provided:

check: allows to check if models are nested. Defalut value is true. Some checkers are not implemented now.

isnested: true when models are checked as nested (manually or automatically). Defalut value is false.

For fitting new models and conducting anova at the same time, see anova_lm for LinearModel, anova_glm for GeneralizedLinearModel, anova_lme for LinearMixedModel, and anova_lfe for FixedEffectModel.

anova(::Type{FTest}, <model>; kwargs...)
anova(::Type{FTest}, <models>...; kwargs...)

Analysis of Variance by F-test.

Return AnovaResult{M, FTest, N}. See AnovaResult for details.

type specifies type of anova:

One LinearModel or GeneralizedLinearModel: 1, 2, 3 are valid

One LinearMixedModel: 1, 3 are valid.

Others: only 1 is valid.

adjust_sigma determines if adjusting to REML when LinearMixedModel is fitted by maximum likelihood.

Note

The result with adjustments will be slightly deviated from that of model fitted directly by REML.

anova(::Type{LRT}, <model>; kwargs...)
anova(::Type{LRT}, <models>...; kwargs...)

Analysis of Variance by likelihood-ratio test.

Return AnovaResult{M, LRT, N}. See AnovaResult for details.

MixedAnova.anova_lm — Function

anova_lm(X, y; test::Type{<: GoodnessOfFit} = FTest, <keyword arguments>) 

anova_lm(test::Type{<: GoodnessOfFit}, X, y; <keyword arguments>)

anova(test::Type{<: GoodnessOfFit}, ::Type{LinearModel}, X, y; 
    type::Int = 1, 
    <keyword arguments>)

ANOVA for simple linear regression.

The arguments X and y can be a Matrix and a Vector or a Formula and a DataFrame.

type specifies type of anova.

dropcollinear controls whether or not lm accepts a model matrix which is less-than-full rank. If true (the default), only the first of each set of linearly-dependent columns

is used. The coefficient for redundant linearly dependent columns is 0.0 and all associated statistics are set to NaN.

anova_lm generate a TableRegressionModel object, which is fitted by lm.

MixedAnova.anova_glm — Function

anova_glm(X, y, d::UnivariateDistribution, l::Link = canonicallink(d); 
        test::Type{<: GoodnessOfFit} = canonicalgoodnessoffit(d), <keyword arguments>)

anova_glm(test::Type{<: GoodnessOfFit}, X, y, d::UnivariateDistribution, l::Link = canonicallink(d); <keyword arguments>)

anova(test::Type{<: GoodnessOfFit}, X, y, d::UnivariateDistribution, l::Link = canonicallink(d); <keyword arguments>)

ANOVA for genaralized linear models.

d: a GLM.UnivariateDistribution.

l: a GLM.Link

For other keyword arguments, see fit.

MixedAnova.anova_lme — Function

anova_lme(f::FormulaTerm, tbl; test::Type{<: GoodnessOfFit} = FTest, <keyword arguments>)

anova_lme(test::Type{<: GoodnessOfFit}, f::FormulaTerm, tbl; <keyword arguments>)

anova(test::Type{<: GoodnessOfFit}, ::Type{<: LinearMixedModel}, f::FormulaTerm, tbl;
        type::Int = 1, 
        adjust_sigma::Bool = true, <keyword arguments>)

ANOVA for linear mixed-effect models.

The arguments f and tbl are Formula and DataFrame.

test: GoodnessOfFit. The default is FTest.

type: type of anova. Only 1, 3 are valid.

adjust_sigma determines whether adjust σ to match that of linear mixed-effect model fitted by REML. The result will be slightly deviated from that of model fitted by REML.

Other keyword arguments

wts = []

contrasts = Dict{Symbol,Any}()

progress::Bool = true

REML::Bool = true

anova_lme generate a LinearMixedModel object which is fitted by lme with REML if applying FTest; otherwise with ML.

MixedAnova.anova_lfe — Function

anova_lfe(f::FormulaTerm, df, vcov::CovarianceEstimator = Vcov.simple(); 
        test::Type{<: GoodnessOfFit} = FTest, <keyword arguments>)
anova_lfe(test::Type{<: GoodnessOfFit}, f::FormulaTerm, df, vcov::CovarianceEstimator = Vcov.simple(); <keyword arguments>)

ANOVA for fixed-effect linear regression.

type: type of anova.

anova_lfe generate a TableRegressionModel{<: FixedEffectModel} object, which is fitted by lfe.

Model fit

MixedAnova.nestedmodels — Function

nestedmodels(trm::TableRegressionModel{<: LinearModel}; null::Bool = false, <keyword arguments>)
nestedmodels(trm::TableRegressionModel{<: GeneralizedLinearModel}; null::Bool = false, <keyword arguments>)
nestedmodels(model::LinearMixedModel; null::Bool = false, <keyword arguments>)

nestedmodels(::Type{LinearModel}, formula, data; null::Bool = true, <keyword arguments>)
nestedmodels(::Type{GeneralizedLinearModel}, formula, data, distr::UnivariateDistribution, link::Link = canonicallink(d); null::Bool = true, <keyword arguments>)
nestedmodels(::Type{LinearMixedModel}, f::FormulaTerm, tbl; null::Bool = true, wts = [], contrasts = Dict{Symbol, Any}(), verbose::Bool = false, REML::Bool = false)

Generate nested models from a model or formula and data.

The null model will be a model with at least one factor (including intercept) if the link function does not allow factors to be 0 (factors in denominators).

InverseLink for Gamma

InverseSquareLink for InverseGaussian

LinearModel fitted with CholeskyPivoted when dropcollinear = true

Otherwise, it will be an empty model.

GLM.glm — Method

GLM.glm(f::FormulaTerm, df::DataFrame, d::Binomial, l::GLM.Link, args...; kwargs...)

Automatically transform dependent variable into 0/1 for family Binomial.

MixedAnova.lme — Function

lme(f::FormulaTerm, tbl; wts, contrasts, progress, REML)

An alias for fit(LinearMixedModel, f, tbl; wts, contrasts, progress, REML).

MixedAnova.glme — Function

glme(f::FormulaTerm, tbl, d::Distribution, l::Link; kwargs...)

An alias for fit(GeneralizedLinearMixedModel, f, tbl, d, l; kwargs...)

MixedAnova.lfe — Function

lfe(formula::FormulaTerm, df, vcov::CovarianceEstimator = Vcov.simple(); kwargs...)

Fit a FixedEffectModel and wrap it into TableRegressionModel.

Warn

This function currently does not perform well. It re-compiles everytime; may be due to @nonspecialize for parameters of reg.

MixedAnova.to_trm — Function

to_trm(model, df)

Wrap fitted FixedEffectModel into TableRegressionModel.

Attributes

MixedAnova.formula — Function

formula(trm::TableRegressionModel)
formula(model::MixedModel)

Unified api for formula of each statistical model.

MixedAnova.anova_test — Function

anova_test(::AnovaResult)

Test statiscics of anova. See AnovaResult for details.

MixedAnova.anova_type — Function

anova_type(aov::AnovaResult)

Type of anova. See AnovaResult for details.

MixedAnova.pval — Function

teststat(aov::AnovaResult)

P-values of test statiscics of anova. See AnovaResult for details.

MixedAnova.teststat — Function

teststat(aov::AnovaResult)

Values of test statiscics of anova. See AnovaResult for details.

StatsAPI.coefnames — Function

coefnames(<model>, anova::Val{:anova})
coefnames(<term>, anova::Val{:anova})

Customize coefnames for anova.

StatsAPI.deviance — Function

deviance(aov::AnovaResult)

Return the stored devaince. The value repressents different statistics for different models and tests.

LinearModel: Sum of Squares.

GeneralizedLinearModel: deviance(model)

MixedModel: NaN when applying FTest; -2loglikelihood(model) == deviance(model) when applying LRT.

When LinearModel is compared to LinearMixedModel, the deviance is alternatively -2loglikelihood(model).

StatsAPI.dof — Function

dof(aov::AnovaResult)

Degree of freedom of models or factors.

StatsAPI.dof_residual — Function

dof_residual(aov::AnovaResult{<: Tuple})
dof_residual(aov::AnovaResult)
dof_residual(aov::AnovaResult{<: MixedModel, FTest})

Degree of freedom of residuals.

Default is applying dof_residual to models in aov.model.

Note

For MixedModels applying FTest, it is calculated by between-within method. See calcdof for details.

StatsAPI.nobs — Function

nobs(aov::AnovaResult{<: Tuple})
nobs(aov::AnovaResult)

Apply nobs to all models in aov.model. See AnovaResult for details.

Miscellaneous

MixedAnova.canonicalgoodnessoffit — Function

canonicalgoodnessoffit(::FixDispDist) = LRT
canonicalgoodnessoffit(::UnivariateDistribution) = FTest

const FixDispDist = Union{Bernoulli, Binomial, Poisson}

Return LRT if the distribution has a fixed dispersion.

MixedAnova.calcdof — Function

calcdof(model::LinearMixedModel)

Calculate degree of freedom of factors and residuals for linear mixed effect models DOF of residuals are estimated by between-within method: dofᵢ = nobsᵢ - dofᵢ₋₁ - nfixᵢ See GLMM FAQ for details.