AnovaMixedModels.jl

AnovaMixedModels.anova_lme

AnovaMixedModels.calcdof

AnovaMixedModels.glme

AnovaMixedModels.lme

ANOVA

Missing docstring.

Missing docstring for AnovaMixedModels.anova(::Val{:AnovaMixedModels}). Check Documenter's build log for details.

AnovaMixedModels.anova_lme — Function

anova_lme(f::FormulaTerm, tbl; test::Type{<: GoodnessOfFit} = FTest, <keyword arguments>)

anova_lme(test::Type{<: GoodnessOfFit}, f::FormulaTerm, tbl; <keyword arguments>)

anova(test::Type{<: GoodnessOfFit}, ::Type{<: LinearMixedModel}, f::FormulaTerm, tbl;
        type::Int = 1, 
        adjust_sigma::Bool = true, <keyword arguments>)

ANOVA for linear mixed-effect models.

The arguments f and tbl are Formula and DataFrame.

test: GoodnessOfFit. The default is FTest.

type: type of anova (1 or 3). Default value is 1.

adjust_sigma: whether adjust σ to match that of linear mixed-effect model fitted by REML. The result will be slightly deviated from that of model fitted by REML.

Other keyword arguments

wts = []

contrasts = Dict{Symbol,Any}()

progress::Bool = true

REML::Bool = true

anova_lme generate a LinearMixedModel fitted with REML if applying FTest; otherwise, a model fitted with ML.

Note

The result with adjust_sigma will be slightly deviated from that of model fitted directly by REML.

Model fit

Missing docstring.

Missing docstring for AnovaMixedModels.nestedmodels(::Val{:AnovaMixedModels}). Check Documenter's build log for details.

AnovaMixedModels.lme — Function

lme(f::FormulaTerm, tbl; wts, contrasts, progress, REML)

An alias for fit(LinearMixedModel, f, tbl; wts, contrasts, progress, REML).

AnovaMixedModels.glme — Function

glme(f::FormulaTerm, tbl, d::Distribution, l::Link; kwargs...)

An alias for fit(GeneralizedLinearMixedModel, f, tbl, d, l; kwargs...)

Miscellaneous

AnovaMixedModels.calcdof — Function

calcdof(model::LinearMixedModel)

Calculate degree of freedom of factors and residuals for linear mixed effect models DOF of residuals are estimated by between-within method: dofᵢ = nobsᵢ - dofᵢ₋₁ - nfixᵢ

Reference algorithm: GLMM FAQ for details.