AnovaBase.jl

ANOVA

AnovaBase.AnovaResult — Type

AnovaResult{M, T, N}

Returned object of anova.

M is a subtype of Tuple if multiple models are provided; otherwise, a type of statistical model.

T is a subtype of GoodnessOfFit; either FTest or LRT.

N is the length of parameters.

Fields

model: full model or a tuple of tested models.

type: type of anova.

dof: degrees of freedom of models or predictors.

deviance: deviance(s) for calculating test statistics. See deviance for more details.

teststat: value(s) of test statiscics.

pval: p-value(s) of test statiscics.

tests: NamedTuple contained extra statistics.

AnovaBase.doc_anova — Constant

anova(<models>...; test::Type{<: GoodnessOfFit}, <keyword arguments>)
anova(Test::Type{FTest}, <model>; <keyword arguments>)
anova(Test::Type{FTest}, <models>...; <keyword arguments>)
anova(Test::Type{LRT}, <model>; <keyword arguments>)
anova(Test::Type{LRT}, <models>...; <keyword arguments>)

Analysis of variance.

Return AnovaResult{M, Test, N}. See AnovaResult for details.

models: model objects. If mutiple models are provided, they should be nested, fitted with the same data and the last one is the most complex.

Test: test statistics for goodness of fit. Available tests are LikelihoodRatioTest (LRT) and FTest.

Models

AnovaBase.doc_nestedmodels — Constant

nestedmodels(<model>; <keyword arguments>)
nestedmodels(<model type>, formula, data; <keyword arguments>)

Generate nested models from a model or formula and data.

Attributes

Missing docstring.

Missing docstring for AnovaBase.formula. Check Documenter's build log for details.

AnovaBase.anova_test — Function

anova_test(::AnovaResult)

Test statiscics of anova. See AnovaResult for details.

AnovaBase.anova_type — Function

anova_type(aov::AnovaResult)

Type of anova. See AnovaResult for details.

AnovaBase.pval — Function

teststat(aov::AnovaResult)

P-values of test statiscics of anova. See AnovaResult for details.

AnovaBase.teststat — Function

teststat(aov::AnovaResult)

Values of test statiscics of anova. See AnovaResult for details.

StatsAPI.deviance — Function

deviance(aov::AnovaResult)

Return the stored devaince. The value repressents different statistics for different models and tests. It may be deviance, Δdeviance, -2loglikelihood or other measures of model performance.

StatsAPI.dof — Method

dof(aov::AnovaResult)

Degrees of freedom of each models or predictors.

StatsAPI.dof_residual — Function

dof_residual(aov::AnovaResult)    
dof_residual(aov::AnovaResult{<: Tuple})

Degrees of freedom of residuals.

By default, it applies dof_residual to models in aov.model.

StatsAPI.nobs — Function

nobs(aov::AnovaResult)
nobs(aov::AnovaResult{<: Tuple})

Number of observations.

Goodness of fit

AnovaBase.GoodnessOfFit — Type

abstract type GoodnessOfFit end

An abstract type as super type of goodness of fit.

AnovaBase.FTest — Type

struct FTest <: GoodnessOfFit end

Type indicates conducting ANOVA by F-test. It can be the first argument or keyword argument test.

AnovaBase.LikelihoodRatioTest — Type

struct LikelihoodRatioTest <: GoodnessOfFit end
const LRT = LikelihoodRatioTest

Type indicates conducting ANOVA by likelihood-ratio test. It can be the first argument or keyword argument test.

AnovaBase.canonicalgoodnessoffit — Function

canonicalgoodnessoffit(::FixDispDist) = LRT
canonicalgoodnessoffit(::UnivariateDistribution) = FTest

const FixDispDist = Union{Bernoulli, Binomial, Poisson}

Return LRT if the distribution has a fixed dispersion; otherwise, FTest.

Developer utility

AnovaBase.ftest_nested — Function

ftest_nested(models::NTuple{N, RegressionModel}, df, dfr, dev, σ²) where N

Calculate F-statiscics and p-values based on given parameters.

models: nested models

df: degrees of freedoms of each models

dfr: degrees of freedom of residuals of each models

dev: deviances of each models, i.e. unit deviance

σ²: squared dispersion of each models

F-statiscic is (devᵢ - devᵢ₋₁) / (dfᵢ₋₁ - dfᵢ) / σ² for the ith predictor.

AnovaBase.lrt_nested — Function

lrt_nested(models::NTuple{N, RegressionModel}, df, dev, σ²) where N

Calculate likelihood ratio and p-values based on given parameters.

models: nested models

df: degrees of freedom of each models

dev: deviances of each models, i.e. unit deviance

σ²: squared dispersion of each models

The likelihood ratio of the ith predictor is LRᵢ = (devᵢ - devᵢ₋₁) / σ².

If dev is alternatively -2loglikelihood, σ² should be set to 1.

AnovaBase.dof_asgn — Method

dof_asgn(v::Vector{Int})

Calculate degrees of freedom of each predictors. 'v' can be obtained by StatsModels.asgn(f::FormulaTerm). For a given trm::RegressionModel, it is as same as trm.mm.assign.

AnovaBase.getterms — Function

getterms(<term>)

Return the symbol of term(s) as a vector of Expr or Symbol.

Examples

julia> iris = dataset("datasets", "iris");

julia> f = formula(lm(@formula(log(SepalLength) ~ Species + PetalLength * PetalWidth), iris))
FormulaTerm
Response:
  (SepalLength)->log(SepalLength)
Predictors:
  1
  Species(DummyCoding:3→2)
  PetalLength(continuous)
  PetalWidth(continuous)
  PetalLength(continuous) & PetalWidth(continuous)

julia> getterms(f)
(Expr[:(log(SepalLength))], [:Species, :PetalLength, :PetalWidth])

julia> getterms(InterceptTerm{true}())
Symbol[]

AnovaBase.isinteract — Function

isinteract(f::MatrixTerm, id1::Int, id2::Int)

Determine if f.terms[id2] is an interaction term of f.terms[id1] and other terms.

Examples

julia> iris = dataset("datasets", "iris");

julia> f = formula(lm(@formula(log(SepalLength) ~ Species + PetalLength * PetalWidth), iris))
FormulaTerm
Response:
  (SepalLength)->log(SepalLength)
Predictors:
  1
  Species(DummyCoding:3→2)
  PetalLength(continuous)
  PetalWidth(continuous)
  PetalLength(continuous) & PetalWidth(continuous)

julia> isinteract(f.rhs, 1, 2)
true

julia> isinteract(f.rhs, 3, 4)
false

julia> isinteract(f.rhs, 4, 5)
true

AnovaBase.doc_select_interaction — Constant

select_super_interaction(f::MatrixTerm, id::Int)
select_sub_interaction(f::MatrixTerm, id::Int)
select_not_super_interaction(f::MatrixTerm, id::Int)
select_not_interaction(f::MatrixTerm, id::Int)

Return a set of index of f.terms, which

return terms are interaction terms of f.terms[id] and other terms.

f.terms[id] is an interaction term of return terms and other terms.

return terms not interaction terms of f.terms[id] and other terms.

f.terms[id] is not interaction term of return terms and other terms.

Examples

julia> iris = dataset("datasets", "iris");

julia> f = formula(lm(@formula(log(SepalLength) ~ Species + PetalLength * PetalWidth), iris))
FormulaTerm
Response:
  (SepalLength)->log(SepalLength)
Predictors:
  1
  Species(DummyCoding:3→2)
  PetalLength(continuous)
  PetalWidth(continuous)
  PetalLength(continuous) & PetalWidth(continuous)

julia> select_super_interaction(f.rhs, 3)
Set{Int64} with 2 elements:
  5
  3

julia> select_sub_interaction(f.rhs, 3)
Set{Int64} with 2 elements:
  3
  1

julia> select_not_super_interaction(f.rhs, 3)
Set{Int64} with 3 elements:
  4
  2
  1

julia> select_not_sub_interaction(f.rhs, 3)
Set{Int64} with 3 elements:
  5
  4
  2

AnovaBase.subformula — Function

subformula(f::FormulaTerm, id; kwargs...)
subformula(lhs::AbstractTerm, rhs::MatrixTerm, id::Int; reschema::Bool = false)
subformula(lhs::AbstractTerm, rhs::MatrixTerm, id; reschema::Bool = false)
subformula(lhs::AbstractTerm, rhs::NTuple{N, AbstractTerm}, id::Int; rhs_id::Int = 1, reschema::Bool = false)

Create formula from existing lhs and rhs (or rhs[tuple_id]) truncated to 1:id or excluded collection id. When id is 0, all terms in rhs (or rhs[tuple_id]) will be removed.

If reschema is true, all terms' schema will be removed.

Examples

julia> iris = dataset("datasets", "iris");

julia> f = formula(lm(@formula(log(SepalLength) ~ Species + PetalLength * PetalWidth), iris))
FormulaTerm
Response:
  (SepalLength)->log(SepalLength)
Predictors:
  1
  Species(DummyCoding:3→2)
  PetalLength(continuous)
  PetalWidth(continuous)
  PetalLength(continuous) & PetalWidth(continuous)

julia> subformula(f, 2)
FormulaTerm
Response:
  (SepalLength)->log(SepalLength)
Predictors:
  1
  Species(DummyCoding:3→2)

julia> subformula(f, [3, 5]; reschema = true)
FormulaTerm
Response:
  (SepalLength)->log(SepalLength)
Predictors:
  1
  Species(DummyCoding:3→2)
  PetalWidth(unknown)

julia> f = formula(fit(LinearMixedModel, @formula(SepalLength ~ SepalWidth + (SepalWidth|Species)), iris))
FormulaTerm
Response:
  SepalLength(continuous)
Predictors:
  1
  SepalWidth(continuous)
  (1 + SepalWidth | Species)

julia> subformula(f, 0)
FormulaTerm
Response:
  SepalLength(continuous)
Predictors:
  0
  (1 + SepalWidth | Species)

AnovaBase.clear_schema — Function

clear_schema(<terms with schema>) = <terms without schema>

Clear any applied schema on terms.

AnovaBase.extract_contrasts — Function

extract_contrasts(f::FormulaTerm)

Extract a dictionary of contrasts. The keys are symbols of term; the values are contrasts (AbstractContrasts).

AnovaBase._diff — Function

_diff(t::NTuple)

Return a tuple of difference between adjacent elements of a tuple(later - former).

AnovaBase._diffn — Function

_diff(t::NTuple)

Return a tuple of difference between adjacent elements of a tuple(former - later).

IO interface

AnovaBase.anovatable — Function

anovatable(aov::AnovaResult{<: RegressionModel})
anovatable(aov::AnovaResult{<: Tuple}, modeltype1, modeltype2)

Return a table with coefficients and related statistics of ANOVA. For nested models, the function will dispatch on the types of the first and the last models. For a single model, no default api was implemented.

The returned AnovaTable object implements the Tables.jl (https://github.com/JuliaData/Tables.jl/) interface, and can be converted e.g. to a DataFrame via using DataFrames; DataFrame(anovatable(aov)).

AnovaBase.add_prednames! — Function

add_prednames!(at::AnovaTable, predictors::Vector{<: AbstractString})
add_prednames!(at::AnovaTable, ::Nothing)

Adding predictors' name to a AnovaTable.

AnovaBase.prednames — Function

prednames(aov::AnovaResult)
prednames(<model>)

Return the name of predictors as a vector of strings. When there are multiple models, return value is nothing.

prednames(<term>)

Return the name(s) of predictor(s). Return value is either a Tuple, a String, an iterable of Strings or nothing.

Examples

julia> iris = dataset("datasets", "iris");

julia> f = formula(lm(@formula(log(SepalLength) ~ SepalWidth + PetalLength * PetalWidth), iris))
FormulaTerm
Response:
  (SepalLength)->log(SepalLength)
Predictors:
  1
  SepalWidth(continuous)
  PetalLength(continuous)
  PetalWidth(continuous)
  PetalLength(continuous) & PetalWidth(continuous)

julia> prednames(f)
("log(SepalLength)", ["(Intercept)", "SepalWidth", "PetalLength", "PetalWidth", "PetalLength & PetalWidth"])

julia> prednames(InterceptTerm{false}())

AnovaBase.testname — Function

testname(::Type{FTest}) = "F test"
testname(::Type{LRT}) = "Likelihood-ratio test"

Name of tests.