AnovaBase.jl

AnovaBase.AnovaResult

AnovaBase.FTest

AnovaBase.GoodnessOfFit

AnovaBase.LRT

AnovaBase.LikelihoodRatioTest

AnovaBase.NoQuote

AnovaBase.OtherStat

AnovaBase.PValue

AnovaBase.TestStat

AnovaBase._diff

AnovaBase._diffn

AnovaBase.anova

AnovaBase.anova

AnovaBase.anova

AnovaBase.anova

AnovaBase.anova_test

AnovaBase.anova_type

AnovaBase.anovatable

AnovaBase.canonicalgoodnessoffit

AnovaBase.clearschema

AnovaBase.extract_contrasts

AnovaBase.formula

AnovaBase.ftest_nested

AnovaBase.getterms

AnovaBase.isinteract

AnovaBase.lrt_nested

AnovaBase.nestedmodels

AnovaBase.nestedmodels

AnovaBase.nestedmodels

AnovaBase.pval

AnovaBase.select_not_sub_interaction

AnovaBase.select_not_super_interaction

AnovaBase.select_sub_interaction

AnovaBase.select_super_interaction

AnovaBase.subformula

AnovaBase.teststat

Type

AnovaBase.AnovaResult — Type

AnovaResult{M, T, N}

Returned object of anova.

M is a subtype of Tuple if multiple models are provided; otherwise, a typeof model.

T is a subtype of GoodnessOfFit; either FTest or LRT.

N is the length of parameters.

Fields

model: full model or tuple of tested models.

type: type of anova.

dof: degree of freedoms of models or factors.

deviance: deviance(s) for calculating test statistics. See deviance for more details.

teststat: value(s) of test statiscics.

pval: p-value(s) of test statiscics.

tests: NamedTuple contained extra statistics.

AnovaBase.FTest — Type

struct FTest <: GoodnessOfFit end

ANOVA by F-test. It can be the first argument or keyword argument test.

AnovaBase.GoodnessOfFit — Type

abstract type GoodnessOfFit end

AnovaBase.LRT — Type

const LRT = LikelihoodRatioTest

See LikelihoodRatioTest.

AnovaBase.LikelihoodRatioTest — Type

struct LikelihoodRatioTest <: GoodnessOfFit end
const LRT = LikelihoodRatioTest

ANOVA by likelihood-ratio test. It can be the first argument or keyword argument test.

ANOVA

AnovaBase.anova — Function

anova(<models>...; test::Type{<: GoodnessOfFit})
anova(test::Type{FTest}, <model>; <keyword arguments>)
anova(test::Type{FTest}, <models>...; <keyword arguments>)
anova(test::Type{LRT}, <model>; <keyword arguments>)
anova(test::Type{LRT}, <models>...; <keyword arguments>)

Analysis of variance.

Return AnovaResult{M, test, N}. See AnovaResult for details.

models: model objects. If mutiple models are provided, they should be nested and the last one is the most complex.

test: test statistics for goodness of fit. Available tests are LikelihoodRatioTest (LRT) and FTest.

Model fit

AnovaBase.nestedmodels — Function

nestedmodels(<model>; <keyword arguments>)
nestedmodels(<model type>, formula, data, <keyword arguments>)

Generate nested models from a model or formula and data.

Attributes

AnovaBase.formula — Function

formula(model)
formula(trm::TableRegressionModel)

Unified api for formula of each statistical model. For a trm::TableRegressionModel, formula is trm.mf.f.

AnovaBase.anova_test — Function

anova_test(::AnovaResult)

Test statiscics of anova. See AnovaResult for details.

AnovaBase.anova_type — Function

anova_type(aov::AnovaResult)

Type of anova. See AnovaResult for details.

AnovaBase.pval — Function

teststat(aov::AnovaResult)

P-values of test statiscics of anova. See AnovaResult for details.

AnovaBase.teststat — Function

teststat(aov::AnovaResult)

Values of test statiscics of anova. See AnovaResult for details.

StatsAPI.coefnames — Function

coefnames(aov::AnovaResult)
coefnames(<model>, anova::Val{:anova})

Customize coefnames for ANOVA. The default method for RegressionModel is applying coefnames on the right hand side of formula.

coefnames(<term>, anova::Val{:anova})

Customize coefnames for ANOVA.

StatsAPI.deviance — Function

deviance(aov::AnovaResult)

Return the stored devaince. The value repressents different statistics for different models and tests. It may be deviance, Δdeviance, -2loglikelihood or other measures of model performance.

StatsAPI.dof — Method

dof(aov::AnovaResult)

Degree of freedom of models or factors.

StatsAPI.dof_residual — Function

dof_residual(aov::AnovaResult{<: Tuple})
dof_residual(aov::AnovaResult)
dof_residual(aov::AnovaResult{<: MixedModel, FTest})

Degree of freedom of residuals.

Default is applying dof_residual to models in aov.model.

StatsAPI.nobs — Function

nobs(aov::AnovaResult{<: Tuple})
nobs(aov::AnovaResult)

Apply nobs to all models in aov.model. See AnovaResult for details.

Developer utility

AnovaBase.ftest_nested — Function

ftest_nested(models::NTuple{N, RegressionModel}, df, dfr, dev, σ²) where N

Calculate F-statiscics and p-values based on given parameters.

models: nested models

df: degrees of freedoms of each models

dfr: degrees of freedom of residuals of each models

dev: deviances of each models

σ²: squared dispersion of each models

F-statiscic is (devᵢ - devᵢ₋₁) / (dfᵢ₋₁ - dfᵢ) / σ² for the ith factor.

AnovaBase.lrt_nested — Function

lrt_nested(models::NTuple{N, RegressionModel}, df, dev, σ²) where N

Calculate likelihood ratio and p-values based on given parameters.

models: nested models

df: degrees of freedoms of each models

dev: deviances of each models

σ²: squared dispersion of each models

The likelihood ratio of the ith factor is LRᵢ = (devᵢ - devᵢ₋₁) / σ².

If dev is alternatively -2loglikelihood, σ² should be 1.

StatsAPI.dof — Method

dof(v::Vector{Int})

Calculate degrees of freedom of each predictors. For a given trm::RegressionModel, v is from trm.mm.assign and must be a non-decreasing array of integers.

AnovaBase.canonicalgoodnessoffit — Function

canonicalgoodnessoffit(::FixDispDist) = LRT
canonicalgoodnessoffit(::UnivariateDistribution) = FTest

const FixDispDist = Union{Bernoulli, Binomial, Poisson}

Return LRT if the distribution has a fixed dispersion.

AnovaBase.isinteract — Function

isinteract(f::MatrixTerm, id1::Int, id2::Int)

Determine if f.terms[id2] is an interaction term of f.terms[id1] and other terms.

AnovaBase.select_super_interaction — Function

select_super_interaction(f::MatrixTerm, id::Int)

Return a set of index of f.terms which are interaction terms of f.terms[id] and other terms.

AnovaBase.select_not_super_interaction — Function

select_not_super_interaction(f::MatrixTerm, id::Int)

Return a set of index of f.terms which are not interaction terms of f.terms[id] and other terms.

AnovaBase.select_sub_interaction — Function

select_sub_interaction(f::MatrixTerm, id::Int)

Return a set of index of f.terms which f.terms[id] are interaction terms of those terms and other terms.

AnovaBase.select_not_sub_interaction — Function

select_not_interaction(f::MatrixTerm, id::Int)

Return a set of index of f.terms which f.terms[id] are not interaction terms of those terms and other terms.

AnovaBase.subformula — Function

subformula(lhs::AbstractTerm, rhs::MatrixTerm, id::Int; reschema::Bool = false)
subformula(lhs::AbstractTerm, rhs::MatrixTerm, id; reschema::Bool = false)
subformula(lhs::AbstractTerm, rhs::NTuple{N, AbstractTerm}, id::Int)

Create formula from existing lhs and rhs truncated to 1:id or excluded collection id.

AnovaBase.getterms — Function

getterms(<term>)

Get an array of Expr or Symbol of terms.

AnovaBase.clearschema — Function

clearschema(<terms with schema>) = <terms without schema>

Clear any applied schema on terms.

AnovaBase.extract_contrasts — Function

extract_contrasts(f::FormulaTerm)

Extract a dictionary of contrasts.

AnovaBase._diff — Function

_diff(t::NTuple)

Return a tuple of difference between adjacent elements of a tuple(later - former).

AnovaBase._diffn — Function

_diff(t::NTuple)

Return a tuple of difference between adjacent elements of a tuple(former - later).

IO interface

AnovaBase.anovatable — Function

anovatable(aov::AnovaResult{<: RegressionModel}; kwargs...)
anovatable(aov::AnovaResult{<: Tuple}, modeltype1, modeltype2; kwargs...)

Return a table with coefficients and related statistics of ANOVA. For nested models, the function will dispatch on the types of the first and the last models. For a single model, no default api was implemented.

The returned AnovaTable object implements the Tables.jl (https://github.com/JuliaData/Tables.jl/) interface, and can be converted e.g. to a DataFrame via using DataFrames; DataFrame(anovatable(aov)).

AnovaBase.TestStat — Type

Show a test statistic using 2 decimal digits

AnovaBase.PValue — Type

Show a p-value using 6 characters, either using the standard 0.XXXX representation or as <Xe-YY.

AnovaBase.OtherStat — Type

Filter NaN

AnovaBase.NoQuote — Type

Wrap a string so that show omits quotes