AnovaFixedEffectModels.jl

ANOVA

AnovaBase.anova — Method

anova(<models>...; test::Type{<: GoodnessOfFit})
anova(test::Type{<: GoodnessOfFit}, <models>...;  <keyword arguments>)

Analysis of variance.

Return AnovaResult{M, test, N}.

models: model objects

TableRegressionModel{<: FixedEffectModel} fitted by AnovaFixedEffectModels.lfe.

If mutiple models are provided, they should be nested and the last one is the most saturated.

test: test statistics for goodness of fit. The default is based on the model type.

TableRegressionModel{<: FixedEffectModel}: FTest.

Other keyword arguments:

When one model is provided:

type specifies type of anova (1 or 3). Default value is 1.

When multiple models are provided:

check: allows to check if models are nested. Defalut value is true. Some checkers are not implemented now.

isnested: true when models are checked as nested (manually or automatically). Defalut value is false.

Algorithm:

For the ith model, devᵢ is defined as the sum of squared deviance residuals (unit deviance). It is equivalent to the residual sum.

The attribute deviance is Δdevᵢ = devᵢ₋₁ - devᵢ.

F-statistic is then defined as Δdevᵢ/(squared dispersion × degree of freedom).

For type I and III ANOVA, F-statistic is computed directly by the variance-covariance matrix(vcov) of the saturated model; the deviance is calculated backward.

Type I:

First, calculate f as the upper factor of Cholesky factorization of vcov⁻¹ * β.

For a factor that starts at ith row/column of vcov with n degree of freedom, the f-statistic is Σᵢⁱ⁺ⁿ⁻¹ fₖ²/n.

Type III:

For a factor occupying ith to jth row/column of vcov with n degree of freedom, f-statistic is (β[i:j]' * vcov[i:j, i:j]⁻¹ * β[i:j])/n.

For fitting new models and conducting anova at the same time, see anova_lfe for FixedEffectModel.

AnovaFixedEffectModels.anova_lfe — Function

anova_lfe(f::FormulaTerm, df, vcov::CovarianceEstimator = Vcov.simple(); 
        test::Type{<: GoodnessOfFit} = FTest, <keyword arguments>)
anova_lfe(test::Type{<: GoodnessOfFit}, f::FormulaTerm, df, vcov::CovarianceEstimator = Vcov.simple(); <keyword arguments>)

ANOVA for fixed-effect linear regression.

vcov: estimator of covariance matrix.

type: type of anova (1 or 3). Default value is 1.

anova_lfe generate a TableRegressionModel{<: FixedEffectModel}.

Models

AnovaFixedEffectModels.lfe — Function

lfe(formula::FormulaTerm, df, vcov::CovarianceEstimator = Vcov.simple(); kwargs...)

Fit a FixedEffectModel and wrap it into TableRegressionModel.

Warn

This function currently does not perform well. It re-compiles everytime; may be due to @nonspecialize for parameters of reg.

AnovaFixedEffectModels.to_trm — Function

to_trm(model, df)

Wrap fitted FixedEffectModel into TableRegressionModel.