AnovaMixedModels.jl

ANOVA

AnovaBase.anova — Method

anova(<models>...; test::Type{<: GoodnessOfFit}, <keyword arguments>)
anova(test::Type{<: GoodnessOfFit}, <models>...; <keyword arguments>)

Analysis of variance.

Return AnovaResult{M, test, N}.

models: model objects

LinearMixedModel fitted by AnovaMixedModels.lme or fit(LinearMixedModel, ...)

GeneralizedLinearMixedModel fitted by AnovaMixedModels.glme or fit(GeneralizedLinearMixedModel, ...)

If mutiple models are provided, they should be nested and the last one is the most saturated. The first model can also be the corresponding GLM object without random effects.

test: test statistics for goodness of fit. Available tests are LikelihoodRatioTest (LRT) and FTest. The default is based on the model type.

LinearMixedModel: FTest for one model fit; LRT for nested models.

GeneralizedLinearMixedModel: LRT for nested models.

Other keyword arguments:

When one model is provided:

type type of anova (1 or 3). Default value is 1.

adjust_sigma: whether adjust σ to match that of linear mixed-effect model fitted by REML. The result will be slightly deviated from that of model fitted by REML.

When multiple models are provided:

check: allows to check if models are nested. Defalut value is true. Some checkers are not implemented now.

isnested: true when models are checked as nested (manually or automatically). Defalut value is false.

Algorithm:

F-test:

No deviance is computed. F-statistic is computed directly from variance-covariance matrix(vcov).

Type I:

First, calculate f as the upper factor of Cholesky factorization of vcov⁻¹ * β.

Then, for a factor that starts at ith row/column of vcov with n degree of freedom, the f-statistic is Σᵢⁱ⁺ⁿ⁻¹ fₖ²/n.

Type III:

For a factor occupying ith to jth row/column of vcov with n degree of freedom, f-statistic is (β[i:j]' * vcov[i:j, i:j]⁻¹ * β[i:j])/n.

LRT:

The attribute deviance is -2loglikelihood for a linear mixed-effect model; deviance computed by Laplace approximation or n-point adaptive Gauss-Hermite quadrature for a generalized linear mixed-effect model.

For the ith model, the likelihood ratio is defined as devianceᵢ₋₁ - devianceᵢ.

For fitting new models and conducting anova at the same time, see anova_lme for LinearMixedModel.

Note

The result with adjust_sigma will be slightly deviated from that of model fitted directly by REML.

Note

For the computation of degrees of freedom, please see calcdof.

AnovaMixedModels.anova_lme — Function

anova_lme(f::FormulaTerm, tbl; test::Type{<: GoodnessOfFit} = FTest, <keyword arguments>)

anova_lme(test::Type{<: GoodnessOfFit}, f::FormulaTerm, tbl; <keyword arguments>)

anova(test::Type{<: GoodnessOfFit}, ::Type{<: LinearMixedModel}, f::FormulaTerm, tbl;
        type::Int = 1, 
        adjust_sigma::Bool = true, <keyword arguments>)

ANOVA for linear mixed-effect models.

The arguments f and tbl are Formula and DataFrame.

test: GoodnessOfFit. The default is FTest.

type: type of anova (1 or 3). Default value is 1.

adjust_sigma: whether adjust σ to match that of linear mixed-effect model fitted by REML. The result will be slightly deviated from that of model fitted by REML.

Other keyword arguments

wts = []

contrasts = Dict{Symbol,Any}()

progress::Bool = true

REML::Bool = true

anova_lme generate a LinearMixedModel fitted with REML if applying FTest; otherwise, a model fitted with ML.

Note

The result with adjust_sigma will be slightly deviated from that of model fitted directly by REML.

Models

AnovaBase.nestedmodels — Method

nestedmodels(model::LinearMixedModel; null::Bool = true, <keyword arguments>)

nestedmodels(::Type{LinearMixedModel}, f::FormulaTerm, tbl; null::Bool = true, wts = [], contrasts = Dict{Symbol, Any}(), verbose::Bool = false, REML::Bool = false)

Generate nested models from a model or formula and data. The null model will be an empty model if the keyword argument null is true (default).

AnovaMixedModels.lme — Function

lme(f::FormulaTerm, tbl; wts, contrasts, progress, REML)

An alias for fit(LinearMixedModel, f, tbl; wts, contrasts, progress, REML).

AnovaMixedModels.glme — Function

glme(f::FormulaTerm, tbl, d::Distribution, l::Link; kwargs...)

An alias for fit(GeneralizedLinearMixedModel, f, tbl, d, l; kwargs...)

Miscellaneous

AnovaMixedModels.calcdof — Function

calcdof(model::LinearMixedModel)

Calculate degree of freedom of factors and residuals for linear mixed effect models DOF of residuals are estimated by between-within method: dofᵢ = nobsᵢ - dofᵢ₋₁ - nfixᵢ

Reference algorithm: GLMM FAQ for details.